+++ Client/include/message_view.hpp
@@ -7,5 +7,5 @@ struct MessageView
 	std::string Author;
 	std::string Content;
 	std::string ID;
-	size_t ChatID;
+	int ChatID;
 };
